Adding support for calibrators, data dumped to stream FIFO
* Calibrators:
Added support for the external calibrator, and custom firmware
to drive the SPI interface to the calibrator. Also ability to
toggle cald/caldb for triggering internal calibrator.

* Stream FIFO:
Now data readout writes to BRAM in addition to sending data to
stream FIFO. When trying to actually readout data, using the stream
FIFO is way easier, as the state machine can send the data and does
not require interacting with any process on HPS.