A bunch of bit reordering
* byte-flips:
In order to actually write out in the order we want, it appears
that we have to reorder each byte of the configuration register.

* register-flips:
And each register is little-endian within its bitrange, so more
bit reordering. This can be turned off with the `little_endian`
kwarg, but it seems to work, so why do that? We also haven't verified
that every register abides this ordering. Currently assuming all do.