Server now running, spawning emitter.
Attempting to use tcp protocol on emitter, as a publisher
This would also relieve the need to run the draft version
of ZMQ.

Still having issues though.

Main problem is an issue with socket reuse, where
the *inproc* socket is the one that I am having trouble with.
Need to figure out some way around this... I don't want to
have to change the channel name every time the program runs!