Committing software for axi-cal-dac
This commit contains a class, `CalCtrl`, for
controlling the axi-calibrator.

Bundles `CalCtrl` with `VataCtrl` in new sictrl
(silicon control) library.

Likely after upcoming merge, that the vatactrl application will be
depricated in favor of sictrl application, which will include both
the vata software, calibration control software, and higher-level
stuff that's being worked on in the silayer-server-* branches