Adding sdk directory for calctrl program
I am super stuck on why the petalinux-build calctrl, with the exact
same code, does not work most of the time.

I have even swapped the addresses around, so it's not an addressing
thing.

Anyways, this will at least work in case the calibrator is needed
before the problem is fixed.