Slowing down cal dac spi clock
Now it should run at 1 MHz.

To slow it down more: open the "customize block" menu for the
axi cal dac IP, and there should be a clock ratio and a width field.
Increase the clock ratio to slow down the sclk. Make sure that the
width value is large enough (2^width > clock ratio)

Clock speed: 100 MHz / 2 / clock-ratio.

The current setting is clock-ratio = 50.