Fixing issue-89
What was causing the issue:
The configuration axi registers are split between the segment
for writing (data set during a set-config), and the segment for reading
(data set in firmware during a get-config). When checking for a successful
set-config in software, we were reading from the first set of registers,
and not the ones set during a get-config. Doing so made sure that
we were always "successful" in setting the config register