Attempt fix to vth/cal flip
It appears that when we try and set the vthr or cal through the dac, we
set the wrong one (attempting to change vthr changes cal, vice versa).
Side is correct though. I think this means we have to change how
spi_sync is connected to the output ports to the following:

spi_sync[0] -> port A,CS1N (should correspond to A-cal)
spi_sync[1] -> port A,CS0N (should correspond to A-vth)
spi_sync[2] -> port B,CS1N (should correspond to B-cal)
spi_sync[3] -> port B,CS0N (should correspond to B-vth)